+++ api/src/main/java/org/jmisb/api/klv/st0102/universalset/SecurityMetadataUniversalSet.java
@@ -33,10 +33,10 @@ public class SecurityMetadataUniversalSet extends SecurityMetadataMessage
     public SecurityMetadataUniversalSet(byte[] bytes) throws KlvParseException
     {
         // Parse the length field
-        LengthField lengthField = BerDecoder.decodeLengthField(bytes, UniversalLabel.LENGTH, false);
-        int lengthLength = lengthField.getSizeOfLength();
+        BerField lengthField = BerDecoder.decode(bytes, UniversalLabel.LENGTH, false);
+        int lengthLength = lengthField.getLength();
         int offset = UniversalLabel.LENGTH + lengthLength;
-        int valueLength = lengthField.getSizeOfValue();
+        int valueLength = lengthField.getValue();
 
         // Parse fields out of the array
         List<UdsField> fields = UdsParser.parseFields(bytes, offset, valueLength);
@@ -72,7 +72,7 @@ public class SecurityMetadataUniversalSet extends SecurityMetadataMessage
             if (bytes != null && bytes.length > 0)
             {
                 chunks.add(key.getUl().getBytes());
-                chunks.add(BerEncoder.encodeLengthField(bytes.length));
+                chunks.add(BerEncoder.encode(bytes.length));
                 chunks.add(bytes.clone());
             }
         }
@@ -93,7 +93,7 @@ public class SecurityMetadataUniversalSet extends SecurityMetadataMessage
         else
         {
             // Prepend length field into front of the list
-            byte[] lengthField = BerEncoder.encodeLengthField(valueLength);
+            byte[] lengthField = BerEncoder.encode(valueLength);
             chunks.add(0, lengthField);
 
             // Prepend key field (UL) into front of the list